move body out from monitor into agent module
